requirements.txt
pillow-8.3.1